student-management/requirement.txt
fastapi
uvicorn
pymongo
python-jose
passlib[bcrypt]
python-dotenv
bcrypt